# docs/swagger.yaml
basePath: /v1
definitions:
...
  backend-go_internal_core_models.Product:
    properties:
      active:
        type: boolean
      category:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: string
      id_user:
        type: string
      is_deleted:
        type: boolean
      name:
        type: string
      price:
        type: number
      sku:
        type: string
      stock:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/backend-go_internal_core_models.User'
    type: object
  backend-go_internal_core_models.Role:
    properties:
      RoleFeature:
        items:
          $ref: '#/definitions/backend-go_internal_core_models.RoleFeature'
        type: array
      active:
        type: boolean
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_deleted:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
  backend-go_internal_core_models.RoleFeature:
    properties:
      activate:
        type: boolean
      create:
        type: boolean
      delete:
        type: boolean
      id_feature:
        type: string
      id_role:
        type: string
      view:
        type: boolean
    type: object
  backend-go_internal_core_models.User:
    properties:
      Role:
        $ref: '#/definitions/backend-go_internal_core_models.Role'
      active:
        type: boolean
      avatar:
        type: string
      cognito_id:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      document:
        type: string
      email:
        type: string
      id:
        type: string
      id_auth:
        type: string
      id_role:
        type: string
      is_deleted:
        type: boolean
      name:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
...
  internal_app_product.CreateProductDTO:
    properties:
      category:
        type: string
      description:
        type: string
      name:
        type: string
      price:
        type: number
      sku:
        type: string
      stock:
        type: integer
    required:
    - category
    - name
    - price
    - sku
    type: object
  internal_app_product.ProductListAllResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/backend-go_internal_core_models.Product'
        type: array
      total:
        type: integer
    type: object
  internal_app_product.ProductListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/backend-go_internal_core_models.Product'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  internal_app_product.UpdateStatusRequest:
    properties:
      active:
        type: boolean
    type: object
...
host: localhost:8888
info:
  contact:
    email: suporte@swagger.io
    name: Suporte API
    url: http://www.swagger.io/support
  description: API modular em Go com Gin e Swagger.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Backend Go API
  version: "1.0"
paths:
...
  /product:
    get:
      description: Retorna produtos com paginação e filtros.
      parameters:
      - description: 'Número da página (padrão: 1)'
        in: query
        name: page
        type: integer
      - description: 'Itens por página (padrão: 25)'
        in: query
        name: size
        type: integer
      - description: Termo de busca
        in: query
        name: searchWord
        type: string
      - description: 'Campos para busca (ex: name,sku,category)'
        in: query
        name: searchFields
        type: string
      - description: Campo para ordenação
        in: query
        name: orderBy
        type: string
      - description: Direção da ordenação (asc/desc)
        in: query
        name: orderDirection
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_app_product.ProductListResponse'
      security:
      - Bearer: []
      summary: Listar Produtos
      tags:
      - Product
    post:
      consumes:
      - application/json
      description: Cria um novo produto com os dados fornecidos.
      parameters:
      - description: Dados do produto
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_app_product.CreateProductDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/backend-go_internal_core_models.Product'
        "400":
          description: Dados inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Criar Produto
      tags:
      - Product
  /product/{id}:
    delete:
      description: Remove um produto logicamente ou fisicamente baseado no ID.
      parameters:
      - description: ID do Produto
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: produto excluído com sucesso'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Excluir Produto
      tags:
      - Product
    get:
      description: Retorna os detalhes de um produto específico.
      parameters:
      - description: ID do Produto
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend-go_internal_core_models.Product'
        "404":
          description: Produto não encontrado
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Obter Produto
      tags:
      - Product
    put:
      consumes:
      - application/json
      description: Atualiza parcialmente um produto baseado no ID.
      parameters:
      - description: ID do Produto
        in: path
        name: id
        required: true
        type: string
      - description: Campos a serem atualizados
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend-go_internal_core_models.Product'
        "400":
          description: Dados inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Atualizar Produto
      tags:
      - Product
  /product/{id}/status:
    patch:
      consumes:
      - application/json
      description: Define se um produto está ativo ou inativo.
      parameters:
      - description: ID do Produto
        in: path
        name: id
        required: true
        type: string
      - description: Novo status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_app_product.UpdateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: status atualizado com sucesso'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Alterar Status do Produto
      tags:
      - Product
  /product/all:
    get:
      description: Retorna todos os produtos (usado para selects/lookups).
      parameters:
      - description: Termo de busca
        in: query
        name: searchWord
        type: string
      - description: Campos para busca
        in: query
        name: searchFields
        type: string
      - description: Campo para ordenação
        in: query
        name: orderBy
        type: string
      - description: Direção da ordenação
        in: query
        name: orderDirection
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_app_product.ProductListAllResponse'
      security:
      - Bearer: []
      summary: Listar Todos os Produtos
      tags:
      - Product
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"